HBASE-2927: BaseScanner gets stale HRegionInfo in some race cases
Summary:
This issue is a branch of HBASE-2812, as it is a specific fix for disabling
tables in some situations. The issue is that the BaseScanner can get a stale HRI
by the time it checks if the region is OFFLINE, and even in trunk where we do a
double checkAssigned there is no way for us to tell if the region was offlined
in the mean time.

Reviewed and committed to public-trunk

Test Plan:
It was committed to public-trunk by jdcryans

DiffCamp Revision: 147520
Reviewed By: jgray
CC: jgray, hbase@lists
Revert Plan:
OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181355 13f79535-47bb-0310-9956-ffa450edef68