Fix location of gc-hbase.log in clusters; rotate the log on process restart
Summary:
#1. gc-hbase.log is set up to  go to $HBASE_HOME/logs.

This is ok for local HBase setup, but this dir doesn't exist in our cluster
setup.

As a result, the gc output was ending up in the *.out (stdout) file, instead of
gc-hbase.log.

In the cluster setup, the logs live in /usr/local/hadoop/logs/<CELLNAME>-HBASE.
So fixed the config to use the above location.

#2. Also added *.out file like log rotation for the gc-hbase.log. Without this,
every time the master or region server is restarted, we lose the old contents of
gc-hbase.log as java does not append to existing file but creates a new one.
With the change, the old file will get moved to gc-hbase.log.1, .2, etc.

Test Plan:
Pushed to cluster, and made sure gc-hbase.log ended up in
/usr/local/hadoop/HBASE.

And that on on a restart, the old log got rotated out to a .1 file.

DiffCamp Revision: 152433
Reviewed By: pkhaitan
CC: pkhaitan, hbase@lists
Revert Plan:
OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181365 13f79535-47bb-0310-9956-ffa450edef68