Eliminate duplicates, stale versions. Have determined behaviour and storefile ordering
Summary:
Goodbye duplicates, hello consistent ordering.

A way to order multiple cells having same keys. Every StoreFile will have a
timestamp associated with it and when any two cells have same keys, they will be
sorted based on timestamps of the StoreFiles they come from. Therefore, for
duplicate versions of a cell, the version which came in latest will be returned.
We are ensuring that only one version will be returned.

There are two components in this patch. The first one is associated with
KeyValueHeap and it ensures that duplicate cells are ordered correctly. The
second one is in ColumnTracker and it ensures that only one version is allowed
to be returned from StoreScanner.

For all existing files and/or any files which do not have the 'TIMESTAMP' field
in meta, their timestamp will be set to zero which means they will be assumed to
be the 'oldest' file.

Also changed the return codes in ColumnTracker to be consistent with definition
of ScanQueryMatcher.ReturnCode.

Included a fix to sort the list of StoreFiles after compaction.

This patch addresses the issues raised in HBASE-1485, HBASE-2406, HBASE-2649,
HBASE-997.

Discussed design issues with Jonathan and Kannan. Accepted by Jonathan and Ryan
in public review.

Test Plan:
All existing tests are passing and running through this code path.

Added two tests in TestColumnSeeking which use several combination of columns,
rows, values and ensure that only the most recent one is returned. This test
will be useful for testing all kinds of column reseeking, row reseeking, etc.

Added tests to TestFromClientSide to do client side testing.

Enabled testDuplicateVersions() which was present in TestFromClientSide.

Fixed two assert statements in testDuplicateVersions()

DiffCamp Revision: 155188
Reviewed By: kannan
CC: davidrecordon, achao, kannan, pkhaitan, hbase@lists
Tasks:
#318553: Fix the ability to query the right column value after a modification
to a particular timestamp.

Revert Plan:
OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181371 13f79535-47bb-0310-9956-ffa450edef68