HBASE-3081: Log Splitting & Replay - Distinguish between Network IOE and Parsing IOE
Summary:
Originally, if HBase got an IOE from HDFS while splitting or opening a region,
it would abort the operation. The assumption being that this is a network
failure that will likely disappear at a later time or different partition of the
network. However, if HBase gets parsing exceptions, we want to log the problem
and continue opening/splitting the region anyways, because parsing is an
idempotent problem and retries won't fix this issue.

Test Plan:
mvn clean test

DiffCamp Revision: 165921
Reviewed By: kannan
CC: dhruba, kannan
Revert Plan:
OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181396 13f79535-47bb-0310-9956-ffa450edef68