remove incorrect region count in Master UI
Summary:
The master web ui displays the region count twice.

The first is an approx region count based on FS, and this is incorrect because
it (the TableDirFilter() logic) doesn't exclude the .oldlogs directory when
checking for Table directories.

The other region count (towards the bottom) is a count of the active regions
and more helpful.

For checking if regions on disk and online regions are consistent, we are
separately doing routine " hbck" calls.

This also slows master UI page load & healthcheck times. On prod cluster the
production cluster, it is reporting that we have 7000 odd regions on FS, while
in reality we just have 472 regions.]

Removing this code from master.jsp. I checked 90 (open source trunk) and this
has already been removed there.

[Separately, need to check why .oldlogs is growing and not getting cleaned up.]

Test Plan:
Run on dev cluster and make sure we no longer display this info.

DiffCamp Revision: 190079
Reviewed By: kranganathan
Reviewers: kranganathan
CC: kranganathan, hbase@lists
Revert Plan:
OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181457 13f79535-47bb-0310-9956-ffa450edef68