Add ColumnRangeFilter
Summary:
Add ColumnRangeFilter. This filter is used for selecting only those keys
with columns that are between minColumn to maxColumn. For example, if minColumn
is 'an', and maxColumn is 'be', it will pass keys with columns like 'ana',
'bad',
but not keys with columns like 'bed', 'eye'. If minColumn is null, there is no
lower bound. If maxColumn is null, there is no upper bound.

Test Plan:
Tested using the unit test:
1. verified the columns we get are as expected.
2. verified that the desired seek behavior happened as expected.
3. tried to measure performance on the unit test with/without seek
optimization. The runtime in the unit test varies quite a bit.

Reviewed By: kannan
Reviewers: jgray, kannan, pkhemani, kranganathan
CC: jfan, kannan, gqchen, hbase@lists
Revert Plan:
OK

Differential Revision: 223913

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181511 13f79535-47bb-0310-9956-ffa450edef68