Allowing [a-f] in hbase-89 region file names to prepare for UUID-based region naming.
Summary:
A trivial regex change in the region file name pattern. We plan to do a rolling
upgrade to this change first. This should have no side effects, but it will
allow to do a rolling upgrade introducing the new region file naming scheme and
adding cache-on-write and evict-on-close to the hbase-89 branch

Test Plan:
Run unit tests.
Run HBaseTest on a 5-node cluster.

Reviewed By: kannan
Reviewers: kenny, jgray, liyintang, kannan
CC: kenny, mbautin, kannan
Tasks:
#496626: better handling of large sized blooms

Revert Plan:
OK

Differential Revision: 232597

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181516 13f79535-47bb-0310-9956-ffa450edef68