Fixing incorrect ByteBuffer instances created by cache-on-write
Summary:
This fixes ByteBuffer objects with arrayOffset=0 and position=8 that were
inserted into the block cache by the cache-on-write functionality, and rewound
to position=0 in loadBlock, resulting in the IllegalArgumentException crash
described in Task #536041, where the DATABLK* magic record was interpreted as a
(key_length, value_length) pair. With this fix the cached buffer has
arrayOffset=8 and position=0, so it is correctly rewound to the beginning of key
length in loadBlock.

Test Plan:
Run HBaseTest with cacheOnWrite and evictOnClose turned on

Reviewed By: kannan
Reviewers: jgray, kannan
CC: kannan, liyintang, hbase@lists,
Revert Plan:
OK

Differential Revision: 236138

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181526 13f79535-47bb-0310-9956-ffa450edef68