First cut of store based compactions
Summary:
Changed all compaction requests to be per store instead of per region.

Note: this is not complete, wanted to get a diff out and continue working.
CompactSplitThread.java:run() is not completely done the way we want to - wanted
to discuss how we wanted to proceed here before doing it.

Test Plan:
Not tested yet.

Reviewed By: nspiegelberg
Reviewers: jgray, nspiegelberg, aaiyer, kannan
Commenters: kannan
CC: , kranganathan, nspiegelberg, kannan
Differential Revision: 233737

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181527 13f79535-47bb-0310-9956-ffa450edef68