Better handling of "node exists" and "no node" exceptions in the recoverable ZooKeeper
Summary:
In Task#536980 we noticed a race condition where a master got restarted, the
backup master came up, and the original master came up as well, resulting in two
active masters. We traced this back to the RecoverableZooKeeper wrapper, which
assumes that if a node creation operation results in a "node exists" exception,
the operation is successful, while in fact the other client grabbed the
ephemeral node. This fix adds a verification of the node contents and re-throws
the NODEEXISTS KeeperException if the data does not match.

Test Plan:
Start a test cluster. Start two masters. Kill the active master and restart
immediately. Verify that both masters compete for the ZK node, but only one
master becomes active.

Reviewed By: liyintang
Reviewers: liyintang, kannan, kranganathan
CC: hbase@lists, , liyintang
Revert Plan:
OK

Differential Revision: 242017

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181532 13f79535-47bb-0310-9956-ffa450edef68