Title: Change the logging level for NotServingRegionException
Summary:Currently the NotServingRegionException is logged at the error level which creates a lot of noise in the scribe logs. Change the logging level for this to info.

Task ID: #586756

Blame Rev:

Reviewed by: kannan

CC:hbase@lists

Test Plan:Started HBaseTests on a dev cluster. While the tests were running, stopped the regionserver. Started it again after a few minutes. This caused the regions to move among the region servers. Since the client had the old location of the regionserver:region mapping, it tried to contact the old region server. This induced the NotServingRegionException. Checked logging for this exception before and after making the change.
With this change the exception will be logged as an info with the classname and the region that caused the exception.
Revert Plan:

Tags:hbase

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 261736

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181549 13f79535-47bb-0310-9956-ffa450edef68