"IdLock" synchronization primitive that will be used to synchronize on block offsets in HFileReaderV2
Summary:
HFile block read operations are currently synchronized on the block index key
object for the particular block. This is done so that only one client goes to
HDFS to read the block and caches it, and all other clients interested in the
block pick it up from the cache. However, with HFile format v2, the data block
index might be multi-level, so there is not a canonical object to use for
synchronizing clients. The IdLock synchronization primitive allows to
synchronize on arbitrary long numbers, creating temporary objects for them as
necessary.

Test Plan:
New unit test. Load testing as part of HFile format v2 testing.

Reviewed By: kannan
Reviewers: kannan, kranganathan, liyintang
Commenters: nspiegelberg, jgray
CC: , hbase@lists, kannan, nspiegelberg, jgray
Revert Plan:
OK

Differential Revision: 248448

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181558 13f79535-47bb-0310-9956-ffa450edef68