Enhance region_status.rb to report status of a single table.
Summary:
For the purpose of Snapshot/Backup we would wish to wait for
all regions of a table to come up before we verify the backed up table.
For this purpose the region_status.rb script needs to be modified. This
functionality should be of great use in other places as well.

Test Plan:
1) Test on dev cluster.

Reviewed By: nspiegelberg
Reviewers: nspiegelberg, madhuvaidya, aaiyer
CC: hbase@lists, , nspiegelberg
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 268472

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181565 13f79535-47bb-0310-9956-ffa450edef68