Adding RowMutation for being able to do both Delete and Put in MR jobs
Summary:
I have added a union-style class called RowMutation to encapsulate either a
Delete or a Put in MapReduce jobs.
This is going to be useful in gathering up mutations for batch import jobs that
also require deletes to be made available.

I have also added a matching RowMutationSortReducer modified from the initial
PutSortReducer so as to know how to sort these elements in a MR job.

Finally, I added a line for selecting the proper reducer in the
HFileOutputFormatter, due to the rather hacky nature of checking for instances
of input...

Test Plan:
I've implemented a small unit test which creates a MR job with a mapper that
statically outputs a Put and a Delete. The new reducer properly sorts them and
outputs them and the test wil dump to a file using HFileOutputFormat.

Reviewed By: kannan
Reviewers: kannan, jfan, nspiegelberg
Commenters: nspiegelberg, gqchen
CC: kannan, nspiegelberg, gqchen, bogdan
Differential Revision: 267252

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181568 13f79535-47bb-0310-9956-ffa450edef68