Don't rewind on a reseek operation in HFile v2 reader
Summary: Not doing a rewind to the first key in block on a reseek operation in
HFile v2 reader, because reseek implies going forward in the file. This fixes a
performance regression introduced in HFile v2 code. This evaded my performance
tests because it only applies to a case of sequential reads with skipping keys,
but I only did random read and contiguous sequential read tests.
Test Plan: Run unit tests. Run HBaseTest.
Reviewed By: kannan
Reviewers: kannan, nspiegelberg
CC: hbase@lists, , kannan
Revert Plan: OK
Differential Revision: 275483

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181592 13f79535-47bb-0310-9956-ffa450edef68