Don't ignore corrupt StoreFiles when opening a region
Summary: We used to ignore StoreFiles that failed to open, which led to a
situation when only a subset of regions was opened, and HBase did not return
results to clients for the affected set of keys. This change makes sure we
propagate IOExceptions coming from an attempt to open a StoreFile all the way up
to HRegionServer.openRegion, where it will lead to a failure to open the whole
region. This way we can avoid returning corrupt data to the application.
Test Plan: Unit tests. 5-node cluster. Dark launch.
Reviewed By: kannan
Reviewers: kannan, nspiegelberg
Commenters: nspiegelberg, dhruba
CC: hbase@lists, , nspiegelberg, dhruba, kannan, mbautin
Revert Plan: OK
Differential Revision: 277786
Task ID: 624776

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181594 13f79535-47bb-0310-9956-ffa450edef68