Increase sleep times on unit tests for async operations.
Summary:
A few unit tests in TestFromClientSide.java used to sleep for a
while after async operations. Due to the sleep duration being small, at
times the unit tests would fail as a result of the async operations not
completing for the duration of the sleep.

Test Plan: 1) Run unit tests
Reviewed By: kannan
Reviewers: kannan
CC: riley, , hbase@lists, kannan
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 278758

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181598 13f79535-47bb-0310-9956-ffa450edef68