Fix RegionServer to correctly retry the failure of closing a region.
Summary:
The region server had a bug where when a region close failed, the
zookeeper was stuck in the state "REGION_CLOSING" for the specified region.
All subsequent retries of closing the region failed since the state in the
zookeeper suggested the region was already being closed by someone.

This needs to be fixed so that, if the re-try is by the same RS, the zookeeper
state can be ignored and a retry of the close region should be performed.

Test Plan:
1) Unit test.
2) Run all tests under hbase.regionserver.

Reviewed By: nspiegelberg
Reviewers: kranganathan, kannan, nspiegelberg
Commenters: liyintang
CC: , hbase@lists, liyintang, nspiegelberg
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 281691
Task ID: 619745

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181604 13f79535-47bb-0310-9956-ffa450edef68