(HBASE-4078) Move corrupt store files out of main store file directory.
Summary:
During compaction of store files, at times the newly created store
file is corrupt. The current logic to handle this is to keep the store file
around but not change the list of store files served. For example if A, B, C
was
compacted to D and D was corrupt, the regionserver still keeps 'D' around but
maintains the current list of store files as A, B, C. The problem now arises
that when the region moves, the new RS sees A, B, C and D and tries to open 'D'
as well and gets an error. This keeps happening as the region moves around.

The fix for now is to move 'D' out of the main store files directory into a
'tmp' location so that regionservers don't pick it up.

Test Plan:
1) Run TestCompaction, TestCompactSelection.
2) Write a unit test which tests this specific case.

Reviewed By: nspiegelberg
Reviewers: kranganathan, kannan, mbautin, nspiegelberg
Commenters: kannan, dhruba, mbautin, liyintang
CC: , hbase@lists, nspiegelberg, kannan, pritam, dhruba,
mbautin, liyintang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 278958
Task ID: 624621

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181914 13f79535-47bb-0310-9956-ffa450edef68