Preserving locality of HRegions on rolling_restarts
Summary:
I created a new jruby script file to take advantage of the HBaseAdmin to
directly hook into the running cluster and perform operations so as to not hack
at it with bash scripts and kill signals.

Then I added message processing between RSs and Master and holding mechanisms
(in memory) at master's side to keep the regions in unassigned mode and not give
them out to other needy servers.

I also fixed a problem with the load balancing code path interfering with
restarts due to some previous issues in the source that got the RS added to the
list of servers used for load calculation before its actual load was refreshed.

There are quite a couple of potential issues at the moment, but I'm putting
this up for initial overview.

Test Plan:
I ran a loadtest on my cluster with 3 regionservers, then started a
rolling-restart and checked in two browser tabs if the test-table region
distribution was identical before and after.

This should however be turned into some form of unit-test...

Reviewed By: kranganathan
Reviewers: kannan, nspiegelberg, liyintang, kranganathan
Commenters: kannan
CC: hbase@lists, bogdan, kannan, kranganathan
Differential Revision: 279896
Task ID: 619449

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181915 13f79535-47bb-0310-9956-ffa450edef68