Port TaskMonitor & MonitoredTask from open source to hbase-90
Summary:
This is a port of the task monitor developed in the open source branch:
https://issues.apache.org/jira/browse/HBASE-3836

I also port the basic functionality of monitoring HMaster and HRegion with
these new tools.

The immediate purpose of this port is to allow implementation of an hbase "show
processlist" (#620988) in a manner that will not require redundancy when
committing to open source.

Test Plan: Ran all unit tests and ran load tests on Nicolas's dev cluster.
Reviewed By: nspiegelberg
Reviewers: nspiegelberg
CC: hbase@lists, riley, nspiegelberg, lshepard, mattwkelly
Differential Revision: 278835

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181923 13f79535-47bb-0310-9956-ffa450edef68