Modify HBase Load Tester to verify entire data.
Summary:
The HBase Load Tester in the read mode would only verify the
prefix of the data instead of the entire data for a particular row. This
has been modified so as to make the value of any cell in the row to be a
deterministic function of the row key itself. Hence the data put into
the row is randomized with a seed being the row key, but since the seed
is the row key we can always deterministically know what its value
should be when we the row.

Test Plan: 1) Run it for HBase running on a 5 node dev cluster.
Reviewed By: kannan
Reviewers: kannan, nspiegelberg
CC: hbase@lists, pritam, kannan, nspiegelberg
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 300961

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181929 13f79535-47bb-0310-9956-ffa450edef68