Clearly report leaked monitored tasks and aborted tasks in the Task Monitor
Summary:
The current reporting of task leaks and aborted tasks is inaccessible and
sometimes nonexistent. This patch improves reporting to be clear, changes the
color-coding for aborted tasks in the JSP task monitor, and provides a color key
to clearly explain to the user which state each row is in.

Additionally, one source of leaked tasks is patched up in store compactions by
aborting at the proper time rather than leaking the task.

Test Plan: Compiled and ran on dev server.
Reviewed By: nspiegelberg
Reviewers: nspiegelberg, kannan
CC: hbase@lists, nspiegelberg
Differential Revision: 305064

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181934 13f79535-47bb-0310-9956-ffa450edef68