Replace jettison JSON encoder use with jackson for consistency with open source.
Summary: The open source trunk is using the jackson JSON encoder
(http://jackson.codehaus.org/) internally. This appears to be better maintained
and faster than jettison. Both for this reason and for consistency with open
source and my slow query log and show processlist patches, I replaced our few
uses of jettison with jackson.
Test Plan: Compiled and ran both the HLogPrettyPrinter and the
HFilePrettyPrinter in JSON mode, validating the JSON.
Reviewed By: nspiegelberg
Reviewers: nspiegelberg
CC: hbase@lists, nspiegelberg
Differential Revision: 306195

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181937 13f79535-47bb-0310-9956-ffa450edef68