Fix the runtime exception when entrySet() is called in SoftValueSortedMap.
Summary:
Currently SoftValueSortedMap.entrySet() tries to iteraate through the entry set
of the underlying map, and put all the values (SoftValue<K,V>) to a newly
created TreeSet<Entry<K,V>>. Because the entry set of SortedMap is already
sorted, so it's not necessary to have a TreeSet to sort those entries again
upon adding. This gets rid of the runtime class cast exception because it does
not require comparing anymore.

Test Plan:
Without the fix, instantiate a SoftValueSortedMap and put three entries in it,
call entrySet().size() would cause it to throw a runtime ClassCastException;
after the fix, it correctly outputs all entries in the sorted order  without
exceptions.

Reviewed By: liyintang
Reviewers: liyintang, kannan
Commenters: kannan
CC: hbase@lists, liyintang, kannan
Differential Revision: 308206
Task ID: 679712

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181942 13f79535-47bb-0310-9956-ffa450edef68