Fixing a parallel locality check issue
Summary:
Solved the block location map clearing issue, added some extra useful logging
on errors and implemented option passing to the locality checker to override
options on the fly. Also circumvented some more possible NPEs.

Also noticed that despite reporting that we are going through a number of
regions (Regions scanned), we actually have roughly twice as many reported files
we are accessing, so the reporting might be a bit misleading...

Test Plan:
I tested it on DL and everything worked pretty much as expected for the number
of threads used:

1)   202323 ms
2)   92715 ms
3)   70758 ms
4)   51976 ms
5)   45343 ms

Reviewed By: kannan
Reviewers: kannan, liyintang
Commenters: liyintang
CC: hbase@lists, achao, kannan, liyintang, bogdan
Blame Revision: D283312
Differential Revision: 308045

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181943 13f79535-47bb-0310-9956-ffa450edef68