Improve RowCounter to count rows in a specific key range.
Summary:
Currently RowCounter in MR package is a very simple map only job, and this
change should let user specify a key range and count the number of rows in this
range. It's done by specifying as arguments "--range=foo,bar" to the program
loader, which would count between ["foo", "bar") that has any column.

Test Plan:
This change is actually a modification used to test diff D295523. First I
loaded rows using recovery utility with specified key range (feature added in
D295523) and let this counter count the same range specified in recovery, and
check if total num of rows equals to num of rows in range. Then, I tried setting
rowcounter.start.key to the end key of that specified in the previous recovery,
and check if there is indeed no row in range.

Reviewed By: kannan
Reviewers: aaiyer, kannan
CC: madhuvaidya, , itapai, kannan
Differential Revision: 296340
Task ID: 659765

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181947 13f79535-47bb-0310-9956-ffa450edef68