Enable preferred assignment in MiniHBaseCluster for unit tests
Summary:
Previously, there were some unit tests failure caused by preferred assignment.
So this preferred assignment was disabled for MiniHBaseCluster since the diff
https://phabricator.fb.com/D271660 .

It is always a good idea to keep the MiniHBaseCluster as the same behavior as a
real HBase cluster does.
Also I have fully understood why the unit test would fail before.
There are 2 reasons contributing to this problem.

1) When enabling the preferred assignment, the master will hold the region
assignment for its best locality region server for a period of time (1 mins), if
its best locality region server has not checked in master.

But there is a potential bug in the code that it will also hold the region
assignment EVEN there is no best locality region server for this region.

This bug only happens when the HBase client tries to create a new table when
the master is trying to hold regions assignment for its best locality region
server, which caused the client throws RetryExhausted Exception.

So fix the bug in this diff.

2) There are some unit tests which depends on load balancer running.
So let these unit test function wait for the load balancer running.

To sum up, this diff is going to enable to preferred assignment in
MiniHBaseCluster for all unit tests, fix 1 assignment bug and update 2 unit
tests.

Test Plan:
Have run all the unit tests.
Only 1 failed unit test case is failed, but this unit test has been failed
recently. Pretty sure that it is not caused by this diff.

testCloseHRegionRetry(org.apache.hadoop.hbase.regionserver.TestHRegionCloseRetry)

Also I will test them on the dev cluster later.

Reviewers: kannan, kranganathan, pritam, gqchen

Reviewed By: kannan

CC: , hbase@lists, kannan

Differential Revision: 312480

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181952 13f79535-47bb-0310-9956-ffa450edef68