Batch CF Updates on table alters
Summary:
HBaseAdmin and HMasterInterface provide no way to batch alterations to a table.
This results in RPC inefficiency, wasteful extra writes to META and, with the
introduction of online schema change, a much worse problem that all regions have
to be restarted for each separate change in a single table alter.

This patch includes a new ColumnOperation, MultiColumnOperation, that performs
all of the changes requested for each region at the same time, and performs a
single write to META for each region. It also includes necessary reorganization
of the existing ColumnOperations to enable this functionality.

It also exposes new methods to both HBaseAdmin clients and HMasterInterface RPC
clients so that they can make batch calls in order to make use of this batched
update feature.

A few design decisions that could be controversial:
1. Four method prototypes are exposed for each type of operation (add, modify,
delete) in HBaseAdmin, plus two for the batched operation. This is in order to
allow easy per-operation batching and in order to not require the user to
perform String-to-bytes conversions. It also follows the current conventions in
HBaseAdmin, which exposes String equivalents for each non-batched operation.
However, it is a possibility that the confusion of having so many options for
each operation is not worth the added convenience. I'm up for suggestions here.
2. Used three lists (one each for add, delete, modify) rather than some kind of
new "MultiAlter" type. This is also for simplicity to the user, but the tradeoff
is not being able to, say, perform String-to-bytes conversions automatically, or
to reduce the number of required method prototypes. Once again, certainly
interested in suggestions!

Test Plan:
Ran all unit tests. No failures.

Basic sanity tests on dev server:
Prior to Patch, for N changes in a single alter, we see N updates, complete
with N region restarts:
hbase(main):003:0> alter 'peeps', {NAME => 'demographics'}, {NAME => 'name',
VERSIONS => '1'}, {NAME => 'location', METHOD => 'delete'}
Updating all regions with the new schema...
0/1 regions updated.
0/1 regions updated.
0/0 regions updated.
Done.
Updating all regions with the new schema...
0/1 regions updated.
0/1 regions updated.
0/0 regions updated.
Done.
Updating all regions with the new schema...
0/1 regions updated.
0/1 regions updated.
0/0 regions updated.
Done.
0 row(s) in 9.2600 seconds

After patch, for N alterations, we have one update:

hbase(main):009:0> alter 'peeps', {NAME => 'demographics', METHOD => 'delete'},
{NAME => 'name', VERSIONS => '2'}, {NAME => 'location', VERSIONS => '1'}
Updating all regions with the new schema...
0/1 regions updated.
0/1 regions updated.
0/0 regions updated.
Done.
0 row(s) in 3.0920 seconds

Reviewers: nspiegelberg, kannan, kranganathan

Reviewed By: kannan

CC: hbase@lists, kannan

Differential Revision: 314294

Task ID: 678576

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181958 13f79535-47bb-0310-9956-ffa450edef68