Make hbase-site-custom.xml override the hbase-site.xml
Summary:
The motivation for diff is that we want to override some config change for any
specific cluster easily by just adding the config entries in the
hbase-site-custom.xml for that cluster.

So adding the hbase-site-custom.xml into HBaseConfiguration.

Also merging hbase-default.xml to hbase-site.xml because both of these files
are the default configuration.

Test Plan: Tested on my dev cluster.

Reviewers: kannan, kranganathan, mbautin, nspiegelberg, gqchen

Reviewed By: kannan

CC: , hbase@lists, liyintang, pkhemani, kannan

Differential Revision: 315123

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181961 13f79535-47bb-0310-9956-ffa450edef68