Caching blocks in seekBefore and refactoring HFile reader access in HFileBlockIndex
Summary:
This fixes a couple of long-existing code issues in HFile v2:
- Making seekBefore cache the previous block it has to read when the scanner
happens to be at the first key of a block (this was a performance regression
introduced in HFile v2).
- Fixing the accounting of the number of blocks read for the one-level index
case in HFileBlockIndex.seekToDataBlock if the current block is the same as the
requested block.
- Getting rid of HFileBlock.BasicReader, which was used both by FSReaderV2 and
HFileReaderV2, but the former did not cache blocks (a source of confusion).
- Adding a new interface HFile.CachingBlockReader instead, which is implemented
by HFile readers and passed to HFileBlockIndex.

Test Plan: Unit tests, dev cluster, dark launch

Reviewers: kannan, liyintang

Reviewed By: kannan

CC: hbase-eng@lists, kannan

Differential Revision: 328625

Revert Plan: OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181979 13f79535-47bb-0310-9956-ffa450edef68