HBASE-4433: avoid extra next (potentially a seek) if done with column/row
Summary:
When we are done with the requested column(s) the code still does an extra
next() call before it realizes that it is actually done. This extra next() call
could potentially result in an unnecessary extra block load. This is likely to
be especially bad for CFs where the KVs are large blobs (such as
Snapshot/ThreadSnapshot CF) where each KV may be occupying a block of its own.
So the next() can often load a new unrelated block unnecessarily.

Test Plan: running unit tests right now; writing a custom test to count number
of blocks accessed with blockcache size set to pretty small; will do dev testing
etc.

Reviewers: mbautin, kranganathan, liyintang, nspiegelberg

Reviewed By: kranganathan

CC: hbase-eng@lists, hbase@lists, kannan, mbautin, nspiegelberg, kranganathan

Differential Revision: 327923

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181982 13f79535-47bb-0310-9956-ffa450edef68