Rename "tab" as "tbl"
Summary:
When open sourcing:
HBASE-4689 (Make the table level metrics work with rpc* metrics),
Nicolas had a good comments that the common abbreviation for table is "tbl"
instead of "tab"  (normally, cut out the vowels).

It would be nice to change if we're not dependent on this format yet.

Test Plan: Running all the unit tests now

Reviewers: kannan, mbautin, nspiegelberg

Reviewed By: kannan

CC: hbase-eng@lists, kannan

Differential Revision: 353246

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1202980 13f79535-47bb-0310-9956-ffa450edef68