Fixed HBaseAdmin ambiguous varargs invocations
Summary:
HBaseAdmin had ambiguous uses of varargs parameters which relied on the
compiler to implicitly cast arguments to Object[] for correct operation.
Certain java compilers (1.6.0_26 on OSX) were found to instead wrap the
argument inside a new Object[], breaking HBaseAdmin and the HBase shell.
This change avoids the ambiguous invocations by explicitly casting the
arguments to Object[].

Test Plan:
compiled HBaseAdmin on dev server and OSX and tested the
HBase shell with commands which take a variable number of arguments.

Reviewers: kranganathan

Reviewed By: kranganathan

CC: hbase-eng@lists, kranganathan

Differential Revision: 354288

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1202982 13f79535-47bb-0310-9956-ffa450edef68