Detect master failover when backup master becomes active
Summary: When the backup master becomes primary, it currently uses the cluster
startup vs. master failover flag that was set in the constructor, potentially a
long time before it becomes primary. This makes the newly active master
re-evaluate that flag.

Test Plan: Run unit tests. Do master failover on a dev cluster and ensure that
the new active master correctly identifies it as a failover and not a cluster
startup.

Reviewers: kranganathan, kannan

Reviewed By: kranganathan

CC: hbase-eng@lists, kranganathan

Differential Revision: 355126

Revert Plan: OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1202983 13f79535-47bb-0310-9956-ffa450edef68