Use a random ZK client port in unit tests so we can run them in parallel
Summary:
The hard-coded ZK client port has long been a problem for running HBase test
suite in parallel. The mini ZK cluster should run on a random free port, and
that port should be passed to all parts of the unit tests that need to talk to
the mini cluster. In fact, randomizing the port exposes a lot of places in the
code where a new configuration is instantiated, and as a result the client tries
to talk to the default ZK client port and times out.

The fix for 89-fb is at https://reviews.facebook.net/D255 (this is very similar
to that). The fix for open-source HBase trunk is at
https://reviews.facebook.net/D279. I guess it should be OK to check this into
the internal branch, since a patch for the trunk is available.

Test Plan: Run unit tests. Deploy on a dev cluster.

Reviewers: kannan, liyintang, kranganathan

Reviewed By: liyintang

CC: hbase-eng@lists, davejwatson, liyintang

Differential Revision: 356873

Revert Plan: OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1202994 13f79535-47bb-0310-9956-ffa450edef68