Handle ZK session expired event properly
Summary:
Previously if any zk client session is expired, it will call
Runtime.getRuntime().halt(1);
So the process of  HBase client, RegionServer, Master  and even unit test will
be aborted.

This revision is going to handle the zk session expired properly.
If the Master or RegionServer has zk session expired. It should shutdown itself
in a control mode.

If the unit test has the zk session expired, it only needs to throw the run
time exception.

If the HBase client gets the zk  session expired, it should reconnect to the zk
again. If the retry is failed, the client should close the zk connection. And
then let the application code to handle this event.

Test Plan: Running all the unit tests, dev cluster test and kill test

Reviewers: kannan, mbautin, kranganathan, pkhemani

Reviewed By: pkhemani

CC: hbase-eng@lists, mbautin, kannan, liyintang, pkhemani, kranganathan

Differential Revision: 354613

Task ID: 753171

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1203002 13f79535-47bb-0310-9956-ffa450edef68