[master] FSUtils can return per-host locality fraction
Summary:
HBase-based block placment benefits from having the fractional locality
of each region on each server. This changes adds a way to output that
from FSUtils. The process is the same as is used to get the best host
per region, but saves the locality in a map.

Test Plan:
The HBase locality check tool outputs the same results as before,
fractional locality output inspected manually.

Reviewers: kranganathan, kannan

Reviewed By: kannan

CC: hbase-eng@lists, kannan, cgist

Differential Revision: 364427

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1212710 13f79535-47bb-0310-9956-ffa450edef68