BaseScanner looks for favorednodes hints in META
Summary:
** 89 master only **
The favored nodes used for explicit placement of regions are persisted
as a column in the .META. table. This change reads those columns in the
base scanner and updates the persistent preferred assignments in the
RegionManager as necessary.

Test Plan:
manually enter hints into META and verify that they are passed to the
region manager by inspecting log output

Reviewers: kranganathan, kannan

Reviewed By: kranganathan

CC: hbase-eng@lists, kranganathan

Differential Revision: 375699

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1225312 13f79535-47bb-0310-9956-ffa450edef68