[master] Munkres' optimal assignment algorithm
Summary:
Explicit placement of regions by locality requires an algorithm for
assigning regions to servers given the locality of each region on each
server. This is an example of the assignment problem which aims to
minimize the total cost (negative locality) of all assignments.

Munkres' algorithm is an optimal algorithm for solving the assignment
problem.  It runs in O(n^3) time and requires O(n^2) space. This may
make it prohibitively costly for large clusters, but remains an option
for moderately sized clusters.  This implementation took ~15s to run for
size n=2000 with real-world cost values.

Test Plan: mvn -Dtest=TestMunkresAssignment test

Reviewers: kranganathan, kannan

Reviewed By: kranganathan

CC: hbase-eng@lists, kranganathan

Differential Revision: 371585

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1225315 13f79535-47bb-0310-9956-ffa450edef68