[master] Average load should compensate for ignored servers
Summary:
When computing the average load across region servers for load balancing
purposes, ServerManager ignores all regions hosted by servers which have
regions being held for them (preferred assignments). However, the number
of servers over which the load was being averaged was not adjusted,
causing the average load to change significantly (and incorrectly)
whenever a preferred assignment was made.

This change adjusts the number of servers when computing the average
load, to compensate for the servers which are ignored.

Test Plan:
mvn -Dtest=TestRegionRebalancing test
Use the shell's move_region command on a balanced cluster and observe
that there are no extraneous region movements.

Reviewers: kranganathan, kannan, liyintang

Reviewed By: liyintang

CC: hbase-eng@lists, kranganathan, liyintang

Differential Revision: 375176

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1227393 13f79535-47bb-0310-9956-ffa450edef68