[master] Regions should not have multiple transient preferred assignments
Summary:
Transient preferred assignments affect the way regions are load balanced
across servers, and those assignments should not exist longer than
necessary. If a server has a transient preferred assignment for a region
which is not in transition, that assignment will prevent the region from
receiving other regions without preferred assignments.

Some unexpected behavior which was observed was a region would have
multiple transient preferred assignments, only one of which could be
honored. The unhonored assignment would cause that server to ignore
assignments of other regions without preferred assignments. If that
server was the least-loaded, the ignored regions would not be assigned
to other servers and would remain unassigned until the preferred
assignment timed out.

This change causes any attempt to add a transient preferred assignment
for a region which already has an assignment to be ignored.

Test Plan:
On a cluster with a mix of regions with and without preferred
assignments, observe from log output that duplicate transient preferred
assignments are prevented and that load balancing does not cause some
regions to be unassigned for extended durations.

Reviewers: kranganathan, kannan, nspiegelberg

Reviewed By: nspiegelberg

CC: hbase-eng@lists, kranganathan, nspiegelberg, cgist

Differential Revision: 376980

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1227394 13f79535-47bb-0310-9956-ffa450edef68