[master] Tool for computing explicit region placement
Summary:
This tool checks the locality of each region on each server and computes
the placement of regions on servers which maximizes total locality. The
placement hints are stored as a column in the .META. table. This tool is
run independently from other processes, but may later be integrated with
the master.

Test Plan:
Run on a cluster, verify that the stored hints reflect the locality of
regions.

Reviewers: kranganathan, kannan, liyintang

Reviewed By: kranganathan

CC: hbase-eng@lists, liyintang, kranganathan

Differential Revision: 379089

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1227397 13f79535-47bb-0310-9956-ffa450edef68